backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.11.3
python-multipart==0.0.20
email-validator==2.2.0
